Move custom systems to the test_astbuilder module.
